Feature: create ShowInfo update, delete html 1. ShowInfo crate, delete 프론트 구현

+++ src/main/java/com/example/stagealarm/show/service/ShowInfoService.java
@@ -131,8 +131,12 @@ public class ShowInfoService {
 
     // 공연 정보 업데이트
     @Transactional
-    public void update(Long id, ShowInfoRequestDto dto) {
+    public void update(Long id, ShowInfoRequestDto dto, MultipartFile file) {
         // 관리자 계정인지 아닌지 확인하는 과정 추가로 필요
+        String authorities = facade.getUserEntity().getAuthorities();
+        if (!authorities.equals("ROLE_ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
 
         ShowInfo showInfo = showInfoRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
 
@@ -144,13 +148,17 @@ public class ShowInfoService {
         showInfo.setTitle(dto.getTitle());
         showInfo.setTicketVendor(dto.getTicketVendor());
         // s3service
-        showInfo.setPosterImage("image");
+        showInfo.setPosterImage(s3FileService.uploadIntoS3("/showInfoImg", file));
         showInfo.setPrice(dto.getPrice());
 
     }
 
     @Transactional
     public void delete(Long id) {
+        UserEntity userEntity = facade.getUserEntity();
+        if (!userEntity.getAuthorities().equalsIgnoreCase("ROLE_ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         showInfoRepository.deleteById(id);
     }
